ng-flask-mysql-docker-example/backend/requirements.txt
Flask==1.0.3

flask-cors==3.0.7

mysql-connector-python==8.0.16
